NAV-222 Refactored blueprints structure.
