--- api/main.py
@@ -1,309 +1,0 @@
-import logging
-
-from flask import Blueprint, jsonify, session
-from flask_login import login_required, current_user
-
-import logic
-import model
-from api import error
-from clients import engine_client
-from clients import ckan_client
-
-main_blueprint = Blueprint('main', __name__)
-log = logging.getLogger(__name__)
-
-
-@main_blueprint.route('/')
-def index():
-    return jsonify({"app": "navigator_api"})
-
-
-@main_blueprint.route('/user')
-@login_required
-def user_details():
-    _user_details = session['ckan_user']
-    return jsonify(
-        {
-            "fullname": _user_details["fullname"],
-            "email": _user_details["email"],
-        }
-    )
-
-
-@main_blueprint.route('/datasets')
-@login_required
-def datasets():
-    ckan_user = session['ckan_user']
-    ckan_cli = _get_ckan_client_from_session()
-    dataset_list = ckan_client.fetch_country_estimates_datasets(ckan_cli)
-    orgs = set(ckan_client.fetch_user_organization_ids(ckan_cli, capacity='create_dataset'))
-    collab_datasets = set(
-        ckan_client.fetch_user_collabolator_ids(ckan_cli, ckan_user_id=ckan_user['id'], capacity='editor')
-    )
-    result = []
-    for dataset in dataset_list:
-        if dataset['organization']['id'] in orgs or dataset['id'] in collab_datasets:
-            result.append({
-                "id": dataset['id'],
-                "organizationName": dataset['organization']['title'],
-                "name": dataset["title"]
-            })
-    return jsonify({
-        "datasets": result
-    })
-
-
-@main_blueprint.route('/workflows')
-@login_required
-def workflow_list():
-    workflows = model.get_workflows(user_id=current_user.id)
-    return jsonify({
-        "workflows": [
-            {
-                "id": workflow.dataset_id,
-                "name": workflow.name
-            } for workflow in workflows
-        ]
-    })
-
-
-def get_or_create_workflow(dataset_id, user_id, name=None):
-    workflow = model.get_workflow(dataset_id, user_id)
-    if not workflow:
-        try:
-            decision_engine = engine_client.get_decision_engine(dataset_id, user_id)
-        except engine_client.EngineError:
-            return None
-        workflow = model.Workflow(
-            dataset_id=dataset_id,
-            name=name,
-            user_id=user_id,
-            decision_engine_id=decision_engine['id']
-        )
-        model.db.session.add(workflow)
-        model.db.session.commit()
-
-    return workflow
-
-
-@main_blueprint.route('/workflows/<dataset_id>/state')
-@login_required
-def workflow_state(dataset_id):
-    ckan_cli = _get_ckan_client_from_session()
-    try:
-        dataset = ckan_client.fetch_dataset_details(ckan_cli, dataset_id)
-    except ckan_client.NotFound:
-        return error.not_found(f"Could not find dataset with id: {dataset_id}")
-    workflow = get_or_create_workflow(dataset['id'], current_user.id, name=dataset['title'])
-    if not workflow:
-        return error.error_response(500, f"Couldn't get decision engine details for dataset {dataset_id}")
-    try:
-        engine_decision = engine_client.get_decision(ckan_cli, dataset_id, skip_actions=workflow.skipped_tasks)
-    except engine_client.EngineError as err:
-        return error.error_response(500, f"Engine error: {err}")
-
-    task_statuses_map = {action['id']: action for action in engine_decision["actions"]}
-    task = engine_decision["decision"]
-
-    task_breadcrumbs = list(task_statuses_map.keys())
-    decision_task_id = str(task["id"])
-    task_details = task["content"]
-    task_progress = engine_decision["progress"]
-    skipped_tasks_to_remove = engine_decision.get("removeSkipActions")
-
-    message = logic.workflow_state_message(workflow, task_breadcrumbs, decision_task_id, skipped_tasks_to_remove)
-
-    _update_last_decision_task_id(decision_task_id, workflow)
-    workflow.task_statuses_map = task_statuses_map
-    model.db.session.add(workflow)
-    model.db.session.commit()
-    if skipped_tasks_to_remove:
-        logic.remove_tasks_from_skipped_list(workflow, skipped_tasks_to_remove)
-
-    current_task = logic.compose_task_details(dataset_id, decision_task_id, task_details,
-                                              task_statuses_map[decision_task_id])
-    return jsonify({
-        "id": f"{workflow.id}",
-        "progress": task_progress["progress"],
-        "message": message,
-        "milestones": task_progress["milestones"],
-        "milestoneListFullyResolved": task_progress["milestoneListFullyResolved"],
-        "taskBreadcrumbs": task_breadcrumbs,
-        "currentTask": current_task
-    })
-
-
-@main_blueprint.route('/workflows/<dataset_id>/state', methods=['DELETE'])
-@login_required
-def workflow_delete(dataset_id):
-    workflow = model.get_workflow(dataset_id, current_user.id)
-    if not workflow:
-        return error.error_response(404, f"Workflow for dataset {dataset_id} not found.")
-    model.db.session.delete(workflow)
-    model.db.session.commit()
-    return jsonify({"message": "success"})
-
-
-def _update_last_decision_task_id(decision_task_id, workflow):
-    workflow.last_engine_decision_id = decision_task_id
-    model.db.session.add(workflow)
-    model.db.session.commit()
-
-
-def is_task_skipped(dataset_id, task_id):
-    workflow = model.get_workflow(dataset_id, current_user.id)
-    return task_id in workflow.skipped_tasks
-
-
-@main_blueprint.route('/workflows/<dataset_id>/tasks/<task_id>')
-@login_required
-def workflow_task_details(dataset_id, task_id):
-    workflow = model.get_workflow(dataset_id, current_user.id)
-    if not workflow:
-        return error.not_found(f"Couldn't get workflow for dataset {dataset_id}")
-    try:
-        action = engine_client.get_action(task_id)
-    except engine_client.EngineError:
-        log.exception(f"Failed to get task details {task_id}", exc_info=True)
-        return error.not_found(f"Failed to get task details {task_id}")
-
-    task = logic.compose_task_details(dataset_id, task_id, action['content'], workflow.task_statuses_map[task_id])
-    return jsonify(task)
-
-
-@main_blueprint.route('/workflows/<dataset_id>/tasks/<task_id>/complete', methods=['GET'])
-@login_required
-def workflow_task_complete_get(dataset_id, task_id):
-    is_completed = logic.is_task_completed(dataset_id, task_id)
-    return jsonify({
-        "id": task_id,
-        "completed": is_completed
-    })
-
-
-@main_blueprint.route('/workflows/<dataset_id>/tasks/<task_id>/complete', methods=['POST'])
-@login_required
-def workflow_task_complete(dataset_id, task_id):
-    ckan_cli = _get_ckan_client_from_session()
-    workflow_state = logic.get_workflow_state(ckan_cli, dataset_id)
-    new_state = logic.complete_task(workflow_state, task_id)
-    try:
-        ckan_client.push_workflow_state(ckan_cli, dataset_id, new_state)
-    except ckan_client.NotFound:
-        return error.not_found(f"Dataset {dataset_id} not found")
-    workflow_task_skip_delete(dataset_id, task_id)
-    return jsonify({"message": "success"})
-
-
-@main_blueprint.route('/workflows/<dataset_id>/tasks/<task_id>/complete', methods=['DELETE'])
-@login_required
-def workflow_task_undo_complete(dataset_id, task_id):
-    ckan_cli = _get_ckan_client_from_session()
-    wf_state = logic.get_workflow_state(ckan_cli, dataset_id)
-    try:
-        new_state = logic.uncomplete_task(wf_state, task_id)
-    except logic.LogicError as e:
-        return error.not_found(str(e))
-    try:
-        ckan_client.push_workflow_state(ckan_cli, dataset_id, new_state)
-    except ckan_client.NotFound:
-        return error.not_found(f"Dataset {dataset_id} not found")
-    return jsonify({"message": "success"})
-
-
-@main_blueprint.route('/workflows/<dataset_id>/tasks/<task_id>/skip', methods=['POST'])
-@login_required
-def workflow_task_skip(dataset_id, task_id):
-    workflow = model.get_workflow(dataset_id, current_user.id)
-    if not workflow:
-        return error.not_found(f"Workflow for dataset {dataset_id} not found.")
-    if task_id not in workflow.skipped_tasks:
-        workflow.skipped_tasks = workflow.skipped_tasks + [task_id]
-        model.db.session.add(workflow)
-        model.db.session.commit()
-    return jsonify({"message": "success"})
-
-
-@main_blueprint.route('/workflows/<dataset_id>/tasks/<task_id>/skip', methods=['DELETE'])
-@login_required
-def workflow_task_skip_delete(dataset_id, task_id):
-    workflow = model.get_workflow(dataset_id, current_user.id)
-    if not workflow:
-        return error.not_found(f"Workflow for dataset {dataset_id} not found.")
-    if task_id not in workflow.skipped_tasks:
-        return error.not_found(f"Task {task_id} not found in skipped tasks for workflow {dataset_id}")
-    else:
-        new_skipped_tasks = set(workflow.skipped_tasks)
-        new_skipped_tasks.remove(task_id)
-        workflow.skipped_tasks = new_skipped_tasks
-        model.db.session.add(workflow)
-        model.db.session.commit()
-    return jsonify({"message": "success"})
-
-
-@main_blueprint.route('/workflows/<dataset_id>/tasks', methods=['GET'])
-@login_required
-def workflow_task_list(dataset_id):
-    ckan_cli = _get_ckan_client_from_session()
-    workflow = model.get_workflow(dataset_id, current_user.id)
-    if not workflow:
-        workflow_state(dataset_id)
-        workflow = model.get_workflow(dataset_id, current_user.id)
-    try:
-        task_list = engine_client.get_workflow_tasks(ckan_cli, dataset_id, skip_actions=workflow.skipped_tasks)
-    except engine_client.EngineError as err:
-        return error.error_response(500, f"Engine error: {err}")
-
-    task_list_with_milestones = logic.get_task_list_with_milestones(dataset_id, task_list['actionList'],
-                                                                    task_list['milestones'])
-
-    return jsonify({
-        "progress": task_list['progress'],
-        "fullyResolved": task_list['fullyResolved'],
-        "taskList": task_list_with_milestones
-    })
-
-
-@main_blueprint.route('/workflows/<dataset_id>/milestones/<milestone_id>', methods=['GET'])
-@login_required
-def workflow_milestone_details(dataset_id, milestone_id):
-    return jsonify({
-        "id": "EST-MIL-01-10-M",
-        "progress": 20,
-        "tasks": [
-            {
-                "completed": True,
-                "id": "EST-OVV-01-10-A",
-                "manual": True,
-                "milestoneID": "EST-MIL-01-10-M",
-                "reached": True,
-                "skipped": False,
-                "title": "Welcome to the UNAIDS HIV Estimates Navigator"
-            },
-            {
-                "completed": False,
-                "id": "EST-OVV-01-11-A",
-                "manual": True,
-                "milestoneID": "EST-MIL-01-10-M",
-                "reached": True,
-                "skipped": False,
-                "title": "Navigator tutorial: Skipping tasks"
-            }
-        ],
-        "title": "Navigator Tutorial"
-    })
-
-
-def _get_ckan_client_from_session():
-    ckan_user = session['ckan_user']
-    ckan_cli = ckan_client.init_ckan(apikey=ckan_user['apikey'])
-    return ckan_cli
-
-
-if __name__ == '__main__':
-    from app import create_app
-
-    app = create_app()
-    with app.app_context():
-
-        pass
