+++ api/workflow/milestone/__init__.py
@@ -1,0 +1,1 @@
+from api.workflow.milestone.routes import milestone_bp
