+++ api/workflow/task/__init__.py
@@ -1,0 +1,1 @@
+from api.workflow.task.routes import task_bp
