+++ app.py
@@ -35,11 +35,11 @@ def create_app(config_object=None):
         json_logging.init_flask(enable_json=True)
         json_logging.init_request_instrument(app)
 
-    from api.auth import auth_blueprint
-    app.register_blueprint(auth_blueprint)
+    from api.auth import auth_bp
+    app.register_blueprint(auth_bp)
 
-    from api.main import main_blueprint
-    app.register_blueprint(main_blueprint)
+    from api import api_bp
+    app.register_blueprint(api_bp)
 
     return app
